Magnetize fix
Fixed a dumbass bug where the map would start to change the pitch/yaw IMMEDIATELY instead of when input is available. This should be better, though there's still some bad repeated code going on. Only when the stick isn't held should we be activating the hovering-over-map sequence, with the audio cue and popup windows.